Changed Application to use u32 for dimensions since that is more common once we are using d3d11 and dxgi functions

+++ src/odin/framework/application.odin
@@ -6,15 +6,15 @@ import "vendor:glfw"
 
 Application :: struct {
 	window : glfw.WindowHandle,
-	dimensions : [2]i32,
+	dimensions : [2]u32,
 	title : string,
 }
 
-GetWindowWidth :: proc (app : ^Application) -> (width : i32) {
+GetWindowWidth :: proc (app : ^Application) -> (width : u32) {
 	return app.dimensions.x
 }
 
-GetWindowHeight :: proc (app : ^Application) -> (height : i32) {
+GetWindowHeight :: proc (app : ^Application) -> (height : u32) {
 	return app.dimensions.y
 }
 
@@ -35,19 +35,19 @@ Initialize :: proc (using app : ^Application) -> (ok : b32) {
 
 	primary_monitor := glfw.GetPrimaryMonitor()
 	video_mode := glfw.GetVideoMode(primary_monitor)
-	dimensions.x = i32(f32(video_mode.width) * 0.9)
-	dimensions.y = i32(f32(video_mode.height) * 0.9)
+	dimensions.x = u32(f32(video_mode.width) * 0.9)
+	dimensions.y = u32(f32(video_mode.height) * 0.9)
 
 	glfw.WindowHint(glfw.SCALE_TO_MONITOR, 0)
 	glfw.WindowHint(glfw.CLIENT_API, glfw.NO_API)
-	window = glfw.CreateWindow(dimensions.x, dimensions.y, strings.clone_to_cstring(title), nil, nil)
+	window = glfw.CreateWindow(i32(dimensions.x), i32(dimensions.y), strings.clone_to_cstring(title), nil, nil)
 	if window == nil {
 		fmt.println("GLFW: Unable to create window")
 		return false
 	}
 
-	window_left := video_mode.width / 2 - dimensions.x / 2
-	window_top := video_mode.height / 2 - dimensions.y / 2
+	window_left := video_mode.width / 2 - i32(dimensions.x) / 2
+	window_top := video_mode.height / 2 - i32(dimensions.y) / 2
 	glfw.SetWindowPos(window, window_left, window_top)
 	return true
 }
